frontend/requirements.txt
pandas
streamlit
pillow
requests
